Simple FLAMES game using Tkinter in Python
Let’s create a Gui version of simple FLAMES game. FLAMES is a popular game named after the acronym: Friends, Lovers, Affectionate, Marriage, Enemies, Sibling. This game does not accurately predict whether or not an individual is right for you, but it can be fun to play this with your friends.